Start moving constants into an EPISODE.CK? file
At the moment, a large number of constants, particularly graphics chunk
numbers and sound chunk numbers, are manually switched in a
episode-specific DefineConstants() function. Instead, lets make these
variables. We make the main file EPISODE.CK?, which includes ACTION and
STRINGS.

Still a lot to do: in some cases we need the values at compile time (or
pointers to the values at compile time), which we can't do with these
dynamically parsed ones, so that'll require some thought.