[Incomplete] Initial SDL3 support
Add VL, SD, and IN backends for SDL3. We do this instead of using the
combined SDL1/2 backends for SD and IN, as so many functions have been
renamed. This does present some problems, as otherwise they share a lot
of code, so one will likely end up outdated.

Otherwise, this is still incomplete:
- No CMake support.
- Sound is broken with SDL_QueueAudio
- Graphics is probably broken on big-endian.
- The pkg-config bits in the makefile are a bit ugly.
- One day we'll want an SDL3_gpu based VL backend, but the renderer is
  good for now.